Refactor forms and views: Moved classes into separate files and folders
